+++ modules/renderer/src/com/genir/renderer/bridge/commands/GL42.java
@@ -1,7 +1,6 @@
 package com.genir.renderer.bridge.commands;
 
 import com.genir.renderer.bridge.context.Context;
-import com.genir.renderer.bridge.context.TextureManager;
 import com.genir.renderer.bridge.interfaces.GLCommand;
 
 import static com.genir.renderer.bridge.context.ContextManager.getThreadContext;