+++ changelog.md
@@ -1,6 +1,7 @@
 - Reduced lazy texture loading lag.
 - Fixed a bug that occasionally prevented saved games from being loaded.
 - Improved file path handling on Linux and Mac. Implemented in cooperation with wing.
+- Fixed a crash when calling glDeleteTextures with a negative parameter.
 
 v0.8.5rc2